pyproject.toml
[project]
name = "dojo-cli"
version = "0.1.0"
description = "A command line interface wrapper for the pwn.college API"
readme = "README.md"
authors = [
    { name = "hidehiroanto", email = "hidehiro.anto@gmail.com" }
]
requires-python = ">=3.14.3"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cairosvg>=2.8.2",
    "itsdangerous>=2.2.0",
    "paramiko>=4.0.0",
    "python-socketio[client]>=5.16.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scp>=0.15.0",
    "textual-image>=0.8.5",
    "trogon>=0.6.0",
    "typer>=0.21.1",
]

[project.scripts]
dojo = "dojo_cli:app"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"
